# doc/API_Documentation.openapi.yaml
openapi: 3.1.0
...
info:
  title: Core User Management API
  version: 1.0.0-draft
  summary: Reusable OpenAPI 3.1 contract for account lifecycle, authentication, user profile management, MFA, passkeys, admin user operations, and privacy preferences.
  description: |
    Canonical OpenAPI 3.1 contract for the reusable core user-management layer in this repository.

    This spec intentionally excludes the PriceTracker business-domain APIs such as catalog, shops, prices,
    purchases, comparisons, watchlists, alerts, and moderation of price submissions.

    The retained surface is limited to reusable platform concerns:

    - registration, login, logout, refresh, password reset, and MFA completion
    - current-user profile, avatar attachment, and contact management
    - passkeys, TOTP MFA, recovery codes, and security summary
    - account-owned private file handling needed by the core, such as avatars or export packages
    - admin user management and system settings
    - legal document discovery, privacy consent capture, and cookie preferences

    Contract conventions:

    - IDs use `format: uuid`
    - timestamps use `format: date-time` in UTC
    - default currency uses ISO 4217 alpha-3 codes
    - response bodies use `camelCase`
    - successful responses use `{ data, meta }`
    - error responses use `{ error }` and include `urgencyLevel` on a 1-9 scale
    - request DTOs reject unknown fields unless explicitly stated otherwise
servers:
...
    description: Production
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      operationId: register
      summary: Register a new user account
      description: Create a standard user account, capture required legal document acceptances, and issue an authenticated session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              default:
                value:
                  email: alex@example.com
                  password: StrongPassword!234
                  displayName: Alex
                  primaryPhone: '+447700900123'
                  acceptedLegalDocuments:
                    - documentKey: terms_of_service
                      version: '2026-04-01'
                    - documentKey: privacy_policy
                      version: '2026-04-01'
      responses:
        '201':
          description: Account created and authenticated session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/register-admin:
    post:
      tags:
        - Auth
      operationId: registerAdminBootstrap
      summary: Register an initial admin account
      description: Setup-only bootstrap route. Keep it behind a deployment flag and disable it outside controlled setup flows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Admin account created and authenticated session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Auth
      operationId: login
      summary: Authenticate with email and password
      description: Authenticate a user. If MFA is enabled, the server returns a challenge instead of a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              default:
                value:
                  email: alex@example.com
                  password: StrongPassword!234
                  rememberMe: true
      responses:
        '200':
          description: Authenticated session returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionEnvelope'
        '202':
          description: Second factor required before a session can be issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaChallengeEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Auth
      operationId: refreshSession
      summary: Refresh an access token
      description: Renew the short-lived access token using the current refresh cookie and CSRF token.
      security:
        - refreshTokenCookie: []
          csrfHeader: []
      responses:
        '200':
          description: Access token refreshed and cookie renewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Auth
      operationId: logout
      summary: Invalidate the current session
      description: Invalidate the active session and expire the session cookie when one exists.
      security:
        - bearerAuth: []
        - refreshTokenCookie: []
          csrfHeader: []
      responses:
        '204':
          description: Session invalidated
  /auth/password/change:
    post:
      tags:
        - Auth
      operationId: changePassword
      summary: Change the current password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/password/forgot:
    post:
      tags:
        - Auth
      operationId: startPasswordReset
      summary: Start the password reset flow
      description: Start a reset flow for a login-enabled email address. Clients should display a neutral confirmation message whether or not the account exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordForgotRequest'
      responses:
        '202':
          description: Reset flow accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/password/reset:
    post:
      tags:
        - Auth
      operationId: completePasswordReset
      summary: Complete the password reset flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '204':
          description: Password reset completed
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/email/verify:
    post:
      tags:
        - Auth
      operationId: verifyEmailChallenge
      summary: Verify an email challenge and activate the related email
      description: Verify a registration or email-change challenge. This endpoint is intended for email-link or emailed-code completion flows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationConfirmRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/email/resend:
    post:
      tags:
        - Auth
      operationId: resendPrimaryEmailVerification
      summary: Resend a primary email verification challenge
      description: Start or resend a primary-account email verification challenge for registration or re-verification flows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationResendRequest'
      responses:
        '202':
          description: Verification challenge accepted for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/mfa/verify:
    post:
      tags:
        - Auth
      operationId: verifyMfaChallenge
      summary: Verify an MFA challenge
      description: Complete a login challenge using TOTP or a recovery code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          description: MFA verified and authenticated session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/passkeys/authentication/options:
    post:
      tags:
        - IdentitySecurity
      operationId: createPasskeyAuthenticationOptions
      summary: Create a WebAuthn authentication challenge
      description: Create a passkey authentication ceremony payload for the browser.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasskeyAuthenticationOptionsRequest'
      responses:
        '200':
          description: WebAuthn authentication options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasskeyAuthenticationOptionsEnvelope'
  /auth/passkeys/authentication/verify:
    post:
      tags:
        - IdentitySecurity
      operationId: verifyPasskeyAuthentication
      summary: Verify a WebAuthn authentication result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasskeyAuthenticationVerifyRequest'
      responses:
        '200':
          description: Passkey verified and authenticated session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /files/uploads:
    post:
      tags:
        - Files
      operationId: createFileUploadIntent
      summary: Create a signed private file upload intent
      description: Create a signed upload intent for account-owned private files such as user avatars or generated export packages.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileUploadIntentRequest'
      responses:
        '201':
          description: Upload intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadIntentEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /files/uploads/{fileId}/complete:
    post:
      tags:
        - Files
      operationId: completeFileUpload
      summary: Finalize a signed file upload
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileIdPath'
      responses:
        '200':
          description: File metadata after successful upload completion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileRecordEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/files/{fileId}:
    get:
      tags:
        - Files
      operationId: getOwnFile
      summary: Get metadata for one caller-owned file
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileIdPath'
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileRecordEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/files/{fileId}/download:
    get:
      tags:
        - Files
      operationId: getOwnFileDownload
      summary: Get a short-lived signed download URL
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileIdPath'
      responses:
        '200':
          description: Signed download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDownloadEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
      example: idem_20260402_001
...
    FileIdPath:
      name: fileId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: One or more fields are invalid.
                  urgencyLevel: 3
                  details:
                    field: email
                  requestId: req_01HZZZZZZZZZZ
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication is required.
                  urgencyLevel: 4
                  requestId: req_01HZZZZZZZZZZ
    Forbidden:
      description: The caller lacks permission for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: FORBIDDEN
                  message: You do not have permission to perform this action.
                  urgencyLevel: 5
                  requestId: req_01HZZZZZZZZZZ
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: NOT_FOUND
                  message: Resource not found.
                  urgencyLevel: 2
                  requestId: req_01HZZZZZZZZZZ
    Conflict:
      description: The request conflicts with existing state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: CONFLICT
                  message: The requested change conflicts with existing state.
                  urgencyLevel: 5
                  requestId: req_01HZZZZZZZZZZ
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: RATE_LIMITED
                  message: Too many requests. Try again later.
                  urgencyLevel: 6
                  requestId: req_01HZZZZZZZZZZ
...
  schemas:
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required:
        - requestId
      properties:
        requestId:
          type: string
          example: req_01HZZZZZZZZZZ
        nextCursor:
          type:
            - string
            - 'null'
          example: null
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
...
    AcknowledgementEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Acknowledgement'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    LegalDocumentAcceptance:
      type: object
      additionalProperties: false
      required:
        - documentKey
        - version
      properties:
        documentKey:
          type: string
          enum:
            - terms_of_service
            - privacy_policy
            - cookie_policy
          example: privacy_policy
        version:
          type: string
          example: '2026-04-01'
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
        - displayName
        - acceptedLegalDocuments
      properties:
        email:
          type: string
          format: email
          example: alex@example.com
        password:
          type: string
          minLength: 12
          maxLength: 128
          example: StrongPassword!234
        displayName:
          type: string
          minLength: 2
          maxLength: 80
          example: Alex
        primaryPhone:
          type:
            - string
            - 'null'
          pattern: ^\+[1-9]\d{6,14}$
          example: '+447700900123'
        acceptedLegalDocuments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LegalDocumentAcceptance'
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: alex@example.com
        password:
          type: string
          minLength: 1
          maxLength: 128
          example: StrongPassword!234
        rememberMe:
          type: boolean
          example: true
...
    AuthSessionEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/AuthSession'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          minLength: 1
          maxLength: 128
          example: OldPassword!234
        newPassword:
          type: string
          minLength: 12
          maxLength: 128
          example: NewPassword!234
    PasswordForgotRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: alex@example.com
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required:
        - resetToken
        - newPassword
      properties:
        resetToken:
          type: string
          minLength: 20
          maxLength: 512
          example: reset_tok_abc123
        newPassword:
          type: string
          minLength: 12
          maxLength: 128
          example: FreshPassword!234
...
    MfaChallengeEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/MfaChallenge'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    MfaVerifyRequest:
      type: object
      additionalProperties: false
      required:
        - challengeId
        - factorType
      properties:
        challengeId:
          $ref: '#/components/schemas/Identifier'
        factorType:
          type: string
          enum:
            - totp
            - recovery_code
          example: totp
        code:
          type:
            - string
            - 'null'
          minLength: 6
          maxLength: 64
          example: '483920'
...
    PasskeyAuthenticationOptionsRequest:
      type: object
      additionalProperties: false
      properties:
        email:
          type:
            - string
            - 'null'
          format: email
          example: alex@example.com
...
    PasskeyAuthenticationOptionsEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/PasskeyAuthenticationOptions'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PasskeyAuthenticationVerifyRequest:
      type: object
      additionalProperties: false
      required:
        - authenticationId
        - credential
      properties:
        authenticationId:
          $ref: '#/components/schemas/Identifier'
        credential:
          type: object
          additionalProperties: true
          example:
            id: credential-id
            rawId: credential-id
            response:
              clientDataJSON: '...'
              authenticatorData: '...'
              signature: '...'
              userHandle: '...'
            type: public-key
...
    FileUploadIntentRequest:
      type: object
      additionalProperties: false
      required:
        - filename
        - contentType
        - size
        - purpose
      properties:
        filename:
          type: string
          minLength: 1
          maxLength: 255
          example: avatar.png
        contentType:
          type: string
          example: image/png
        size:
          type: integer
          minimum: 1
          maximum: 10485760
          example: 231231
        purpose:
          $ref: '#/components/schemas/FilePurpose'
        checksumSha256:
          type:
            - string
            - 'null'
          pattern: ^[A-Fa-f0-9]{64}$
          example: 3f786850e387550fdab836ed7e6dc881de23001b
...
    FileUploadIntentEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/FileUploadIntent'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    FileRecordEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/FileRecord'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    FileDownloadEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/FileDownload'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    EmailVerificationConfirmRequest:
      type: object
      additionalProperties: false
      required:
        - verificationToken
      properties:
        verificationToken:
          type: string
          minLength: 20
          maxLength: 512
          example: verify_tok_abc123
    EmailVerificationResendRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: alex@example.com
        purpose:
          type:
            - string
            - 'null'
          enum:
            - registration
            - reverification
            - null
          example: registration
...